Mostly fix build issues for arm target
pdd builds (and works), but still need to straighten out some things in pd-test
