+++ devices/Cargo.toml
@@ -5,14 +5,14 @@ authors = ["Max Willsey <me@mwillsey.com>"]
 edition = "2018"
 
 [dependencies]
-rppal = "0.11.2"
-structopt = "0.2.15"
-log = "0.4.6"
-env_logger = "0.6.1"
-signal-hook = "0.1.9"
-
 config = "0.9"
+env_logger = "0.6.1"
+log = "0.4.6"
+rppal = "0.11.2"
 serde = "1"
+signal-hook = "0.1.9"
+structopt = "0.3.5"
+toml = "0.5"
 
 [build-dependencies]
 chrono = "0.4"
