+++ devices/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod error;
+pub mod hv507;
+pub mod max31865;
+pub mod mcp4725;
+pub mod pca9685;
