+++ purpledrop/src/lib.rs
@@ -5,6 +5,7 @@ pub mod error;
 pub mod board;
 pub mod httpserver;
 pub mod location;
+pub mod purpledrop;
 #[cfg(not(target_arch="arm"))]
 pub mod mockpurpledrop;
 pub mod purpledrop_rpc;
